Automatically discover parameters when user supplies an entire class (#34)
When authoring a Task in an assembly, task parameters are declared
simply by creating public properties. CodeTaskFactory instead requires
the user to manually declare all properties in a <ParameterGroup/>
element. This is tedious when using Type="Class" because the
parameters must be listed both as property declarations and in
<ParametersGroup/>. Also, this requirement increases maintenance when
using inheritence to extend or intercept an existing Task from another
assembly.

This change causes simple reflection to be used to calculate the parameters
for all Type="Class" tasks. A warning is emitted if <ParameterGroup/> is
specified and it is ignored.

Fixes #33.